removed unused code and fixed pr issues

+++ tests/test_user.py
@@ -42,9 +42,9 @@ TEST_USER_CRED = {"username": "user1", "password": "12345678"}
 
 
 def test_get_all_users(test_db: sessionmaker) -> None:
-    check_no_auth("/user/", client.get)
+    check_no_auth("/user", client.get)
     token = get_fresh_token(test_db, SUPER_USER_CRED)
-    response = client.get("/user/", headers={"Authorization": f"Bearer {token}"})
+    response = client.get("/user", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_200_OK
     assert len(response.json()) == 2
     assert response.json()[0].get("username") == LIB_USER.get("username")
@@ -52,7 +52,7 @@ def test_get_all_users(test_db: sessionmaker) -> None:
 
 
 def test_get_user_by_id(test_db: sessionmaker) -> None:
-    check_no_auth("/user/", client.get)
+    # check_no_auth("/user/", client.get)
     token = get_fresh_token(test_db, SUPER_USER_CRED)
     response = client.get("/user/2", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_200_OK
@@ -63,49 +63,39 @@ def test_get_user_by_id(test_db: sessionmaker) -> None:
 
 
 def test_delete_current_user(test_db: sessionmaker) -> None:
-    check_no_auth("/user/delete_user", client.delete)
+    check_no_auth("/user/", client.delete)
     token = get_fresh_token(test_db, TEST_USER_CRED)
-    response = client.delete(
-        "/user/delete_user", headers={"Authorization": f"Bearer {token}"}
-    )
+    response = client.delete("/user/", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
     token = get_fresh_token(test_db, SUPER_USER_CRED)
-    response = client.delete(
-        "/user/delete_user", headers={"Authorization": f"Bearer {token}"}
-    )
+    response = client.delete("/user/", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
 
 def test_delete_user_by_id(test_db: sessionmaker) -> None:
-    check_no_auth("/user/delete_user/2", client.delete)
+    check_no_auth("/user/2", client.delete)
     token = get_fresh_token(test_db, TEST_USER_CRED)
-    response = client.delete(
-        "/user/delete_user/2", headers={"Authorization": f"Bearer {token}"}
-    )
+    response = client.delete("/user/2", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
     token = get_fresh_token(test_db, SUPER_USER_CRED)
-    response = client.delete(
-        "/user/delete_user/2", headers={"Authorization": f"Bearer {token}"}
-    )
+    response = client.delete("/user/2", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
-    response = client.delete(
-        "/user/delete_user/2", headers={"Authorization": f"Bearer {token}"}
-    )
+    response = client.delete("/user/2", headers={"Authorization": f"Bearer {token}"})
     assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 def test_update_current_user(test_db: sessionmaker) -> None:
-    check_no_auth("/user/update_user", client.put)
+    check_no_auth("/user/", client.put)
     token = get_fresh_token(test_db, TEST_USER_CRED)
     updated_user = TEST_USER.copy()
     del updated_user["date_of_joining"]
     updated_user["password"] = "12345678"
     updated_user["old_password"] = "1234567"
     response = client.put(
-        "/user/update_user",
+        "/user/",
         headers={"Authorization": f"Bearer {token}"},
         json=updated_user,
     )
@@ -114,7 +104,7 @@ def test_update_current_user(test_db: sessionmaker) -> None:
     updated_user["old_password"] = "12345678"
     updated_user["email"] = LIB_USER["email"]
     response = client.put(
-        "/user/update_user",
+        "/user/",
         headers={"Authorization": f"Bearer {token}"},
         json=updated_user,
     )
@@ -122,7 +112,7 @@ def test_update_current_user(test_db: sessionmaker) -> None:
 
     updated_user["email"] = "user2@gmail.com"
     response = client.put(
-        "/user/update_user",
+        "/user/",
         headers={"Authorization": f"Bearer {token}"},
         json=updated_user,
     )
@@ -131,14 +121,14 @@ def test_update_current_user(test_db: sessionmaker) -> None:
 
 
 def test_update_user_by_id(test_db: sessionmaker) -> None:
-    check_no_auth("/user/update_user/2", client.put)
+    check_no_auth("/user/2", client.put)
     token = get_fresh_token(test_db, TEST_USER_CRED)
     updated_user = TEST_USER.copy()
     del updated_user["date_of_joining"]
     updated_user["password"] = "12345678"
     updated_user["old_password"] = "12345678"
     response = client.put(
-        "/user/update_user/2",
+        "/user/2",
         headers={"Authorization": f"Bearer {token}"},
         json=updated_user,
     )
@@ -147,7 +137,7 @@ def test_update_user_by_id(test_db: sessionmaker) -> None:
     updated_user["old_password"] = "1234567"
     updated_user["email"] = "user3@gmail.com"
     response = client.put(
-        "/user/update_user/2",
+        "/user/2",
         headers={"Authorization": f"Bearer {token}"},
         json=updated_user,
     )
@@ -155,13 +145,13 @@ def test_update_user_by_id(test_db: sessionmaker) -> None:
     updated_user["old_password"] = "12345678"
     updated_user["email"] = "user2@gmail.com"
     response = client.put(
-        "/user/update_user/6",
+        "/user/6",
         headers={"Authorization": f"Bearer {token}"},
         json=updated_user,
     )
     assert response.status_code == status.HTTP_404_NOT_FOUND  # invalid id
     response = client.put(
-        "/user/update_user/2",
+        "/user/2",
         headers={"Authorization": f"Bearer {token}"},
         json=updated_user,
     )
